Create packages for java files
