+++ data/chad_data.txt
@@ -1,4 +1,2 @@
 T | 0 | write
 T | 0 | seat
-T | 0 | s
-T | 0 | sssa
