+++ src/main/java/chad/Chad.java
@@ -1,5 +1,9 @@
-import java.io.IOException;
+package chad;
 
+import chad.exception.ChadException;
+import chad.task.Task;
+
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Scanner;
