+++ src/main/java/chad/task/Todo.java
@@ -1,8 +1,10 @@
+package chad.task;
+
 /**
- * Todo task that is child of task
+ * chad.task.Todo task that is child of task
  */
 public class Todo extends Task {
-    Todo(String description) {
+    public Todo(String description) {
         super(description);
     }
     @Override
